Created a folder for footer stuff and changed coloring to the primary variable
