+++ src/components/common/index.js
@@ -4,7 +4,7 @@ import List from './List';
 import LoadingComponent from './LoadingComponent';
 import Button from './Button';
 import Header from './Header';
-import Footer from './Footer';
+import Footer from './Footer/Footer';
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
 // to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
 export {
